Renamed LeaveRegisterScreen file
