+++ lib/employee/utilities/dashboard_item_list.dart
@@ -3,7 +3,7 @@ import 'package:echno_attendance/employee/hr_screens/attendance_report/employee_
 import 'package:echno_attendance/employee/hr_screens/employee_register.dart';
 import 'package:echno_attendance/employee/hr_screens/update_details_screen.dart';
 import 'package:echno_attendance/employee/models/dashboard_item.dart';
-import 'package:echno_attendance/leave_module/screens/leave_register.dart';
+import 'package:echno_attendance/leave_module/screens/leave_register_screen.dart';
 import 'package:flutter/material.dart';
 
 class DashboardItemList {
